+++ TombEngine/Scripting/Internal/TEN/Util/Util.cpp
@@ -189,8 +189,7 @@ namespace TEN::Scripting::Util
 	// end
 	static sol::optional <std::unique_ptr<Moveable>> PickMoveable(const Vec2& screenPos)
 	{
-		auto realScreenPos = PercentToScreen(screenPos.x, screenPos.y);
-		auto ray = GetRayFrom2DPosition(Vector2(int(std::get<0>(realScreenPos)), int(std::get<1>(realScreenPos))));
+		auto ray = GetRayFrom2DPosition(PercentToScreen(screenPos));
 
 		auto vector = Vector3i::Zero;
 		int itemNumber = ObjectOnLOS2(&ray.first, &ray.second, &vector, nullptr);
@@ -216,8 +215,7 @@ namespace TEN::Scripting::Util
 	// end
 	static sol::optional <std::unique_ptr<Static>> PickStatic(const Vec2& screenPos)
 	{
-		auto realScreenPos = PercentToScreen(screenPos.x, screenPos.y);
-		auto ray = GetRayFrom2DPosition(Vector2((int)std::get<0>(realScreenPos), (int)std::get<1>(realScreenPos)));
+		auto ray = GetRayFrom2DPosition(PercentToScreen(screenPos));
 
 		StaticMesh* mesh = nullptr;
 		auto vector = Vector3i::Zero;
@@ -251,24 +249,6 @@ namespace TEN::Scripting::Util
 		return posA.Distance(posB);
 	}
 
-	/// Converts ObjectID to a string. Used by Custom Inventory module to retrieve examine texts.
-	// @function GetObjectIDString
-	// @tparam Objects.ObjID objectID ID of the object.
-	// @treturn string ObjectID converted to string.
-	static std::string GetObjectIDString(GAME_OBJECT_ID objectID)
-	{
-		return GetObjectName(objectID);
-	}
-
-	/// Runs the OnUseItem callback. Used by Custom Inventory module to enable OnItemUse callbacks.
-	// @function OnUseItemCallBack
-	static void OnUseItemCallBack()
-	{
-		g_GameScript->OnUseItem((GAME_OBJECT_ID)g_Gui.GetInventoryItemChosen());
-		HandleAllGlobalEvents(EventType::UseItem, (Activator)short(LaraItem->Index));
-
-	}
-
 	void Register(sol::state* state, sol::table& parent)
 	{
 		auto tableUtil = sol::table(state->lua_state(), sol::create);
@@ -292,8 +272,6 @@ namespace TEN::Scripting::Util
 			)
 		);
 		tableUtil.set_function(ScriptReserved_PrintLog, &PrintLog);
-		tableUtil.set_function(ScriptReserved_GetObjectIDString, &GetObjectIDString);
-		tableUtil.set_function(ScriptReserved_OnUseItemCallBack, &OnUseItemCallBack);
 
 		// COMPATIBILITY
 		tableUtil.set_function("CalculateDistance", &CalculateDistance);