+++ CHANGELOG.md
@@ -46,16 +46,18 @@ TombEngine releases are located in this repository (alongside with Tomb Editor):
 
 ### Lua API changes
 * Added `Collision.Ray` class for raycasting to detect rooms, items, and static meshes.
+* Added `View.DisplayItem` class that can be used to render 3D items in 2D space.
 * Added `View.DisplayAnchors` primitive class and `View.DisplaySprite:GetAnchors` method.
 * Added `UI` section to `Flow.Settings` class with several options to customize main menu and system text.
-* Added `Inventory.Reset` function to reset inventory state to default.
+* Added several `Inventory` functions, including `Inventory.UseItem` and `Inventory.ResetToDefault`.
 * Added `ambientOcclusion` parameter to `Flow.Settings.Graphics` to force-disable ambient occlusion option.
 * Added `killPoisonedEnemies` parameter to `Flow.Settings.Gameplay` to enable or disable killing enemies with poisoned ammo.
 * Added a safeguard against adding or removing callback functions within the same callback type.
 * Added a safeguard for `Moveable:Enable` method called for already destroyed moveables.
+* Added `Moveable:HideInteractionHighlight` method.
 * Added `Strings.DisplayStringOption.VERTICAL_BOTTOM` flag.
-* Added optional type argument for inventory item and interaction type for `LaraObject.Interact` function.
-* Added optional type argument for interaction type for `Object.ShowInteractionHighlight` function.
+* Added optional type argument for inventory item and interaction type for `LaraObject:Interact` function.
+* Added optional type argument for interaction type for `Object:ShowInteractionHighlight` function.
 * Added `Vec3:Direction` function and `Vec3:Translate` overload.
 * Added direction vector constructor to `Rotation`.
 * Added overloads for `TEN.Util.PercentToScreen(Vec2(x, y))` and `TEN.Util.ScreenToPercent(Vec2(x, y))`.